+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "botmock-rasa-export",
-  "version": "1.1.0",
+  "version": "2.0.0",
   "description": "build rasa bots from botmock projects",
   "scripts": {
     "test": "jest",
@@ -34,11 +34,8 @@
     "@botmock-api/flow": "0.7.0",
     "@botmock-api/log": "0.2.0",
     "@botmock-api/text": "^0.2.2",
-    "@botmock-api/utils": "2.3.1",
-    "chalk": "2.4.2",
-    "dotenv": "8.1.0",
+    "dotenv": "^8.2.0",
     "fs-extra": "8.1.0",
-    "node-fetch": "2.6.0",
     "uuid": "3.3.3",
     "yaml": "1.6.0"
   },
@@ -46,7 +43,6 @@
     "@types/fs-extra": "8.0.0",
     "@types/jest": "24.0.22",
     "@types/node": "12.12.6",
-    "@types/node-fetch": "2.5.2",
     "@types/uuid": "^3.4.6",
     "jest": "24.9.0",
     "ts-jest": "24.1.0",
